+++ src/roguelike_game/ecs/fsm/states/monster/aggro_state.py
@@ -1,6 +1,4 @@
 from roguelike_game.ecs.fsm.state import State
-from roguelike_game.ecs.fsm.states.flee_state import FleeState
-from roguelike_game.ecs.fsm.states.attack_state import AttackState
 from roguelike_game.ecs.fsm.states.death_state import DeathState
 from roguelike_engine.config.config_tiles import TILE_SIZE
 
@@ -26,6 +24,7 @@ class AggroState(State):
         # Verificar salud para cambio a huida
         health_cmp = world.components['Health'][entity]
         if health_cmp.current_hp <= health_cmp.max_hp * 0.3:
+            from roguelike_game.ecs.fsm.states.monster.flee_state import FleeState
             world.components['NPCState'][entity].fsm.change_state(FleeState(), entity)
             return
         # Verificar distancia de ataque
@@ -37,10 +36,11 @@ class AggroState(State):
             dist_sq = dx*dx + dy*dy
             mr_cmp = world.components['MeleeRange'][entity]
             if dist_sq <= (mr_cmp.range * TILE_SIZE) ** 2:
+                from roguelike_game.ecs.fsm.states.attack_state import AttackState
                 world.components['NPCState'][entity].fsm.change_state(AttackState(), entity)
                 return
-        # Si no ataca ni huye, continuar persiguiendo
-        from roguelike_game.ecs.fsm.states.chase_state import ChaseState
+        # Si no ataca ni huye, continuar persiguiendo        
+        from roguelike_game.ecs.fsm.states.monster.chase_state import ChaseState
         world.components['NPCState'][entity].fsm.change_state(ChaseState(), entity)
         return
