+++ src/roguelike_game/ecs/fsm/states/monster/patrol_state.py
@@ -1,6 +1,5 @@
 import math
 from roguelike_game.ecs.fsm.state import State
-from roguelike_game.ecs.fsm.states.idle_state import IdleState
 from roguelike_game.ecs.fsm.states.death_state import DeathState
 from roguelike_game.ecs.components.transform.velocity import Velocity
 from roguelike_engine.config.config_tiles import TILE_SIZE
@@ -34,8 +33,8 @@ class PatrolState(State):
             dx_p = pos.x - player_pos.x
             dy_p = pos.y - player_pos.y
             if dx_p*dx_p + dy_p*dy_p <= (world.components['AggroRange'][eid].radius * TILE_SIZE) ** 2:
+                from roguelike_game.ecs.fsm.states.monster.aggro_state import AggroState
                 npc_state = world.components['NPCState'][eid]
-                from roguelike_game.ecs.fsm.states.aggro_state import AggroState
                 npc_state.fsm.change_state(AggroState(), entity)
                 return
         # Mover hacia el waypoint actual
