+++ src/roguelike_game/ecs/systems/combat/spells/spell_casting_system.py
@@ -5,7 +5,7 @@ y jugadores.
 """
 from roguelike_game.ecs.fsm.states.cast_state import CastState
 from roguelike_game.ecs.systems.fsm.fsm_system import _EntityProxy
-from roguelike_game.ecs.fsm.states.aggro_state import AggroState
+from roguelike_game.ecs.fsm.states.monster.aggro_state import AggroState
 from roguelike_game.ecs.components.fsm.npc_state import NPCState
 from roguelike_game.ecs.fsm.fsm import FiniteStateMachine
 import pygame
